Additional Enlarge Emote and Minor Adjustment to Resource Cog (#72)
